+++ NEWS.md
@@ -2,6 +2,7 @@
   * Feature: lintr checks to vignettes
   * Feature: pkgdown site
   * Feature: Install dependencies by default
+  * Feature: Set PKG_TARBALL in the .Renviron on build step
 
 # Version 0.1
   * Initial version